requirements.txt
azure-cognitiveservices-vision-customvision
pillow==12.0.0
matplotlib
fastapi[all]
uvicorn[standard]
numpy==2.3.3
torch==2.9.1
torchvision==0.24.1
